// client/linup-logic/src/app/services/generated/model/apiSupportTicketsPostRequest.ts
/**
 * API V1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiSupportTicketsPostRequest { 
    reply_to: string;
    message: string;
    /**
     * Leave blank (spam honeypot). Hidden field.
     */
    honeypot?: string;
}